backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]>=3.1
python-decouple==3.8
openpyxl==3.1.5
